# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LizyStudio",
    "description": "Web GUI for LizyML",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/inference/run": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Inference Run",
        "description": "Run inference with a path to data (H-0009).",
        "operationId": "inference_run_api_inference_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Inference Run Api Inference Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inference/upload": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Inference Upload",
        "description": "Upload data file for inference (H-0015).",
        "operationId": "inference_upload_api_inference_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inference_upload_api_inference_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Inference Upload Api Inference Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inference/history": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Inference History",
        "description": "List inference records. job_id optional \u2014 omit for all records (H-0010).",
        "operationId": "inference_history_api_inference_history_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Inference History Api Inference History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inference/{inf_id}": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Inference Get",
        "description": "Get inference record metadata.",
        "operationId": "inference_get_api_inference__inf_id__get",
        "parameters": [
          {
            "name": "inf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inf Id"
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Inference Get Api Inference  Inf Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inference/{inf_id}/predictions": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Inference Predictions",
        "description": "Get paginated predictions table.",
        "operationId": "inference_predictions_api_inference__inf_id__predictions_get",
        "parameters": [
          {
            "name": "inf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inf Id"
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "rows",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Rows"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Inference Predictions Api Inference  Inf Id  Predictions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/inference/{inf_id}/plot/{plot_type}": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Inference Plot",
        "description": "Get a plot from the model used for inference.",
        "operationId": "inference_plot_api_inference__inf_id__plot__plot_type__get",
        "parameters": [
          {
            "name": "inf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inf Id"
            }
          },
          {
            "name": "plot_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plot Type"
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Inference Plot Api Inference  Inf Id  Plot  Plot Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inference/{inf_id}/download": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Inference Download",
        "description": "Download predictions as CSV.",
        "operationId": "inference_download_api_inference__inf_id__download_get",
        "parameters": [
          {
            "name": "inf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inf Id"
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inference/{inf_id}/comparison/{other_inf_id}": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Inference Comparison",
        "description": "Compare two inference runs.",
        "operationId": "inference_comparison_api_inference__inf_id__comparison__other_inf_id__get",
        "parameters": [
          {
            "name": "inf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inf Id"
            }
          },
          {
            "name": "other_inf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Other Inf Id"
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Inference Comparison Api Inference  Inf Id  Comparison  Other Inf Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_inference_upload_api_inference_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_inference_upload_api_inference_upload_post"
      },
      "DataSource": {
        "properties": {
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "path"
        ],
        "title": "DataSource"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RunRequest": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "data": {
            "$ref": "#/components/schemas/DataSource"
          },
          "return_shap": {
            "type": "boolean",
            "title": "Return Shap",
            "default": false
          },
          "evaluate": {
            "type": "boolean",
            "title": "Evaluate",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "data"
        ],
        "title": "RunRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}